[fix] move publish command to publish folder
